B034 The constants to file refactoring
