+++ middleware/celeb-validate-req.js
@@ -2,7 +2,7 @@ const { celebrate, Joi } = require('celebrate');
 Joi.objectId = require('joi-objectid')(Joi);
 
 const { isObjectIdValid, urlValidatorCheck } = require('../helpers/helpers');
-const { errors } = require('../helpers/errorMessages');
+const { errors } = require('../configs/errorMessages');
 
 const validateSignup = celebrate(
   {
